Мелкие ошибки и переименование

+++ src/main/java/take_nft/ru/takeNft/dto/TopHolderDto.java
@@ -4,5 +4,5 @@ public record TopHolderDto(
         String walletId,
         String username,
         String avatarUrl,
-        long nftCount
+        int nftCount
 ) {}
